+++ system/src/Grav/Common/Twig/TwigEnvironment.php
@@ -10,6 +10,9 @@
 namespace Grav\Common\Twig;
 
 use Twig\Environment;
+use Twig\Error\LoaderError;
+use Twig\Template;
+use Twig\TemplateWrapper;
 
 /**
  * Class TwigEnvironment
@@ -18,4 +21,38 @@ use Twig\Environment;
 class TwigEnvironment extends Environment
 {
     use WriteCacheFileTrait;
+
+    /**
+     * @inheritDoc
+     */
+    public function resolveTemplate($names)
+    {
+        if (!\is_array($names)) {
+            $names = [$names];
+        }
+
+        $count = \count($names);
+        foreach ($names as $name) {
+            if ($name instanceof Template) {
+                return $name;
+            }
+            if ($name instanceof TemplateWrapper) {
+                return $name;
+            }
+
+            if (1 !== $count && !$this->getLoader()->exists($name)) {
+                continue;
+            }
+
+            try {
+                return $this->loadTemplate($name);
+            } catch (LoaderError $e) {
+                if (1 === $count) {
+                    throw $e;
+                }
+            }
+        }
+
+        throw new LoaderError(sprintf('Unable to find one of the following templates: "%s".', implode('", "', $names)));
+    }
 }